Fixed verb tense agreement
The verb "have" refers to a plural noun: (parts of a transaction) while other verbs in the sentence (ie, has) refer to the singular noun transaction. To resolve this confusion I have removed the (parts of a transaction) section. This allows subject and verb to be in agreement while preserving the meaning of the sentence.